Some changes for organize API files..
