+++ app/Api/V1/Controllers/Admin/Users/Users_List_Controller.php
@@ -2,8 +2,8 @@
 
 namespace App\Api\V1\Controllers\Admin\Users;
 
-use Illuminate\Http\Request;
-use App\Http\Controllers\Controller;
+use Illuminate\Http\Request,
+    App\Http\Controllers\Controller;
 
 class Users_List_Controller extends Controller
 {
