+++ app/Api/V1/Controllers/Auth/UserController.php
@@ -2,8 +2,8 @@
 
 namespace App\Api\V1\Controllers\Auth;
 
-use App\Api\V1\Controllers\Controller;
-use Auth;
+use App\Api\V1\Controllers\Controller,
+    Auth;
 
 class UserController extends Controller
 {
